Added a PasswordEncrypter program compatible with apache-commons-codec Java library.
Also used UnixCrypt directly for CRYPT-PW scheme instead on relying on Crypt.
This makes WHOIS code compatible with PasswordEncrypter program.

The PasswordEncrypter program should be made available to users of WHOIS for them
to generate MD5-PW and CRYPT-PW for mntner object, auth: attribute.
It is usually made available via a webpage on RIR website.