PGPKEY Modify Issue :
The attribute FINGERPR is saved in a table, however other GENERATED attribute types are not. On update, attributes,those stored in tables, are deleted then reinserted, however the code did not remove all those of GENERATED types, hence the FINGERPR was not deleted, and on insertion, we were having the error duplicate entry for primary key.

Now only those with indexstrategy unindexed are skipped.
Note: PARENT attribute type was missing an indexstrategy, added that as well, so that it does not go through to try deleting it.